questions/urls.py
# questions/urls.py
from django.urls import path
from rest_framework.routers import DefaultRouter
from . import views

router = DefaultRouter()
router.register(r"domains", views.DomainViewSet, basename="domains")
router.register(r"topics", views.TopicViewSet, basename="topics")
router.register(r"", views.QuestionViewSet, basename="questions")

urlpatterns = router.urls
